unified app with sql query replay and readme
