--- experimental-tools/pgrr/requirements.txt
@@ -1,2 +1,0 @@
-click>=8.0.0
-asyncpg==0.29.0
