--- experimental-tools/pgrr/setup.py
@@ -1,32 +1,0 @@
-"""Setup configuration for pgrr CLI tool"""
-
-from setuptools import setup, find_packages
-
-with open("README.md", "r", encoding="utf-8") as fh:
-    long_description = fh.read()
-
-setup(
-    name="pgrr",
-    version="0.1.0",
-    author="UC Berkeley Codebase",
-    author_email="your.email@example.com",
-    description="A powerful CLI tool",
-    long_description=long_description,
-    long_description_content_type="text/markdown",
-    url="https://github.com/yourusername/pgrr",
-    packages=find_packages(),
-    classifiers=[
-        "Programming Language :: Python :: 3",
-        "License :: OSI Approved :: MIT License",
-        "Operating System :: OS Independent",
-    ],
-    python_requires=">=3.8",
-    install_requires=[
-        "click>=8.0.0",
-    ],
-    entry_points={
-        "console_scripts": [
-            "pgrr=pgrr.cli:cli",
-        ],
-    },
-)
